extract legacy pareto and validation modules
